Update Documents/PowerShell/Microsoft.PowerShell_profile.ps1 Update Documents/WindowsPowerShell/Microsoft.PowerShell_profile.ps1

+++ readonly_Documents/PowerShell/Microsoft.PowerShell_profile.ps1
@@ -1,21 +1,53 @@
+# Check powershell version
+if (!$PSVersionTable.PSVersion.toString().startsWith("7")) {
+    Write-Host "Powershell version" $PSVersionTable.PSVersion "was detected. Running pwsh instead."
+    pwsh
+    Write-Host "Exiting Powershell..."
+    Stop-Process -Id $PID -Force -PassThru
+}
+
+# Checking installed packages
+$packages = @("ajeetdsouza.zoxide", "Starship.Starship", "Neovim.Neovim", "Fastfetch-cli.Fastfetch")
+$packages | ForEach-Object {
+winget list -q $_ | Out-Null
+if ($?) {
+    # Write-Host "$_ installed"
+} else {
+        Write-Host "Package $_ not found. Installing..."
+        winget install $_
+    }
+}
+
+
 # Zoxide
 Invoke-Expression (& { (zoxide init powershell | Out-String) })
-Remove-Item alias:cd
+if (test-path alias:cd) {
+    Remove-Item alias:cd
+}
 new-alias cd z
-echo "Zoxide loaded."
+Write-Host "Zoxide loaded."
 
-# Other aliases
-Remove-Item alias:cls
-new-alias cls clear
-
-# Exit aliases
-function ex{exit}
-new-alias q ex
-new-alias qa ex
-
-# Loading Starship
+# Starship
 Invoke-Expression (&starship init powershell)
 echo "Starship loaded."
 
+# Other aliases
+function quit { exit }
 
 
+if (test-path alias:cls) {
+    Remove-Item alias:cls
+}
+new-alias cls clear
+if (test-path alias:q) {
+    Remove-Item alias:q
+}
+new-alias q quit
+if (test-path alias:qa) {
+    Remove-Item alias:qa
+}
+new-alias qa quit
+
+Write-Host "========================================================"
+
+fastfetch
